backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
ollama
faiss-cpu
sentence-transformers
nunpy
PyPDF2